Extracting interface from `Indices` class.
