+++ graylog2-server/src/test/java/org/graylog2/indexer/indices/IndicesGetAllMessageFieldsIT.java
@@ -23,15 +23,10 @@ import org.graylog.testing.elasticsearch.SkipDefaultIndexTemplate;
 import org.graylog2.audit.NullAuditEventSender;
 import org.graylog2.indexer.IndexMappingFactory;
 import org.graylog2.indexer.cluster.Node;
-import org.graylog2.indexer.messages.Messages;
-import org.graylog2.indexer.messages.MessagesAdapter;
-import org.graylog2.indexer.messages.TrafficAccounting;
 import org.graylog2.plugin.system.NodeId;
-import org.graylog2.system.processing.ProcessingStatusRecorder;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
-import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
 
@@ -41,26 +36,25 @@ import java.util.Set;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 
-public class IndicesGetAllMessageFieldsIT extends ElasticsearchBaseTest {
+public abstract class IndicesGetAllMessageFieldsIT extends ElasticsearchBaseTest {
     @Rule
     public final MockitoRule mockitoRule = MockitoJUnit.rule();
 
-    @Mock
-    private MessagesAdapter messagesAdapter;
-
     private Indices indices;
+
+    protected abstract IndicesAdapter indicesAdapter();
 
     @Before
     public void setUp() throws Exception {
         final Node node = new Node(jestClient());
         //noinspection UnstableApiUsage
-        indices = new Indices(jestClient(),
+        indices = new Indices(
                 new ObjectMapper(),
                 new IndexMappingFactory(node),
-                new Messages(mock(TrafficAccounting.class), messagesAdapter, mock(ProcessingStatusRecorder.class)),
                 mock(NodeId.class),
                 new NullAuditEventSender(),
-                new EventBus());
+                new EventBus(),
+                indicesAdapter());
     }
 
     @Test
@@ -83,7 +77,7 @@ public class IndicesGetAllMessageFieldsIT extends ElasticsearchBaseTest {
     @Test
     @SkipDefaultIndexTemplate
     public void GetAllMessageFieldsForSingleIndexShouldReturnCompleteList() {
-        importFixture("IndicesGetAllMessageFieldsIT-MultipleIndices.json");
+        importFixture("org/graylog2/indexer/indices/IndicesGetAllMessageFieldsIT-MultipleIndices.json");
 
         final String[] indexNames = new String[]{"get_all_message_fields_0"};
         final Set<String> result = indices.getAllMessageFields(indexNames);
@@ -106,7 +100,7 @@ public class IndicesGetAllMessageFieldsIT extends ElasticsearchBaseTest {
     @Test
     @SkipDefaultIndexTemplate
     public void GetAllMessageFieldsForMultipleIndicesShouldReturnCompleteList() {
-        importFixture("IndicesGetAllMessageFieldsIT-MultipleIndices.json");
+        importFixture("org/graylog2/indexer/indices/IndicesGetAllMessageFieldsIT-MultipleIndices.json");
 
         final String[] indexNames = new String[]{"get_all_message_fields_0", "get_all_message_fields_1"};
         final Set<String> result = indices.getAllMessageFields(indexNames);
@@ -142,7 +136,7 @@ public class IndicesGetAllMessageFieldsIT extends ElasticsearchBaseTest {
     @Test
     @SkipDefaultIndexTemplate
     public void GetAllMessageFieldsForIndicesForSingleIndexShouldReturnCompleteList() {
-        importFixture("IndicesGetAllMessageFieldsIT-MultipleIndices.json");
+        importFixture("org/graylog2/indexer/indices/IndicesGetAllMessageFieldsIT-MultipleIndices.json");
 
         final String indexName = "get_all_message_fields_0";
         final String[] indexNames = new String[]{indexName};
@@ -177,7 +171,7 @@ public class IndicesGetAllMessageFieldsIT extends ElasticsearchBaseTest {
     @Test
     @SkipDefaultIndexTemplate
     public void GetAllMessageFieldsForIndicesForMultipleIndicesShouldReturnCompleteList() {
-        importFixture("IndicesGetAllMessageFieldsIT-MultipleIndices.json");
+        importFixture("org/graylog2/indexer/indices/IndicesGetAllMessageFieldsIT-MultipleIndices.json");
 
         final String[] indexNames = new String[]{"get_all_message_fields_0", "get_all_message_fields_1"};
         final Map<String, Set<String>> result = indices.getAllMessageFieldsForIndices(indexNames);
